separate nft allowance integration tests
